Extract VaccinatorDropDown into separate file
